Fix grid coord type problems with baked in vertex offsets
If vertex offsets are in use, they are currently expected to be baked in
the grid position query parameter. The main reason for this is how
CATMAID interacts with this library at the moment. Regardless of its
use, the consequence is that grid positions can become negative, which
isn't expected by Rust and causes problems when input values are
negative as they are interpreted by Rust as u64 as well, causing an
undetected overflow. This is fixed now by allowing negative grid
position parameters.